+++ src/app/features/tenant/services/tenant.service.ts
@@ -1,8 +1,8 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import {Observable} from "rxjs";
-import {TenantBusinessModel} from "../business-models/tenant.business-model";
-import {environment} from "../../../environments/environment";
+import {TenantBusinessModel} from "../../../common/business-models/tenant.business-model";
+import {environment} from "../../../../environments/environment";
 
 @Injectable({
   providedIn: 'root'
@@ -14,5 +14,9 @@ export class TenantService {
 
   public getTenants(): Observable<TenantBusinessModel[]> {
     return this.http.get<TenantBusinessModel[]>(`${environment.apiUrl}/tenants`)
+  }
+
+  public postTenant(tenantBusinessModel: TenantBusinessModel): Observable<TenantBusinessModel> {
+    return this.http.post<TenantBusinessModel>(`${environment.apiUrl}/tenants`, tenantBusinessModel)
   }
 }