+++ src/app/features/tenant/services/tenant.service.ts
@@ -27,4 +27,8 @@ export class TenantService {
   public putTenant(tenantBusinessModel: TenantBusinessModel, tenantId: string): Observable<TenantBusinessModel> {
     return this.http.put<TenantBusinessModel>(`${environment.apiUrl}/tenants/${tenantId}`, tenantBusinessModel);
   }
+
+  public deleteTenant(tenantId: string): Observable<TenantBusinessModel> {
+    return this.http.delete<TenantBusinessModel>(`${environment.apiUrl}/tenants/${tenantId}`);
+  }
 }